+++ apm-agent-api/src/main/java/co/elastic/apm/api/TransactionImpl.java
@@ -96,4 +96,10 @@ class TransactionImpl extends AbstractSpanImpl implements Transaction {
         return "";
     }
 
+    @Override
+    public Transaction setStartTimestamp(long epochMicros) {
+        doSetStartTimestamp(epochMicros);
+        return this;
+    }
+
 }